backend/requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
PyJWT==2.10.1
sqlparse==0.5.5
tzdata==2025.3
